Blanks space "gotcha" got ME!
My CSS habits got the best of me, and I had to spend quiiiiiite a while chasing this one down. Thus, I thought it might be worthwhile to save others some time and trouble by roping off this potential pitfall with some kind of a note in the docs. Essentially: "You can't have a blank space after the colon in a partial call!" (Please feel free to choose better wording and/or placement.)