Widget vs. Widgets
The link that works is https://docs.statamic.com/addons/classes/widget (no "s"); https://docs.statamic.com/addons/classes/widgets (with an "s") gives you a 404.

IMHO, that's backwards. You should change the working link to the one WITH the "s" and fix the links to suit. The link-text has it with an "s" in both places on the page—it'd be only natural to go to a likewise "plural" URL, if you know what I mean.