Add available order variable to collection tag
When trying to access the number of a collection's item (where order is set to number and not date), you can use `{{ order }}`. Slightly different than `index` or `zero_index` because it is not in context of the current loop, but of the original collection. I guess it would be similar to `{{ date }}` on a date-based collection.