Deleted a sentence for clarity
The "More information about this is below." from recursive children variable description makes no sense anymore. This information is now above, and it's also the only description that makes a direct connection to a piece of content on the same page. So I think it's not needed and does not add more clarity.