denoting -> noting
Is it the shutting down itself that "denotes"? Otherwise (more likely) did the people who did the shutting down justify it by observing "the fact that nothing was being marked up"? In that case, to "denote" is wrong. To denote is not to observe, claim, or argue -- it is to symbolize. Minimizing my suggestion by just removing that "de".